requirements.txt
pymupdf==1.24.1
pillow==12.1.0
numpy==1.26.4
scikit-learn==1.8.0
matplotlib==3.10.8
opencv-python==4.11.0.86